# eva-proyecto/eva-backend/docs/api/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EVA - Sistema de Gestión de Equipos API",
        "version": "1.0.0",
        "description": "API completa para el sistema de gestión de equipos biomédicos EVA",
        "contact": {
            "email": "admin@eva-system.com",
            "name": "Equipo de Desarrollo EVA"
        }
    },
    "servers": [
        {
            "url": "http:\/\/localhost:8000\/api",
            "description": "Servidor de Desarrollo"
        }
    ],
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Token de autenticación Sanctum"
            }
        }
    },
...
    "paths": {
...
        "\/dashboard\/stats": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Obtener estadísticas del dashboard",
                "description": "Estadísticas generales del sistema",
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estadísticas del dashboard"
                    }
                }
            }
        }
    }
}